Создала класс SingOut2Test.java для US_08.002 написала тест кейс. этот автотест проверяет что разлогинивание происходит моментально и еще удалила класс не привязанный к тест кейсу
